+++ site/src/components/HelpTooltip/HelpTooltip.tsx
@@ -207,7 +207,7 @@ const styles = {
     display: "flex",
     alignItems: "center",
     ...(theme.typography.body2 as CSSObject),
-    color: theme.experimental.roles.active.fill.outline,
+    color: theme.roles.active.fill.outline,
   }),
 
   linkIcon: {
