+++ site/src/modules/dashboard/DeploymentBanner/DeploymentBannerView.tsx
@@ -379,7 +379,7 @@ const HealthIssue: FC<PropsWithChildren> = ({ children }) => {
     <Stack direction="row" spacing={1} alignItems="center">
       <ErrorIcon
         css={{ width: 16, height: 16 }}
-        htmlColor={theme.experimental.roles.error.outline}
+        htmlColor={theme.roles.error.outline}
       />
       {children}
     </Stack>
