+++ site/src/pages/TemplatePage/TemplateInsightsPage/TemplateInsightsPage.tsx
@@ -420,10 +420,7 @@ const TemplateUsagePanel: FC<TemplateUsagePanelProps> = ({
   const totalInSeconds =
     validUsage?.reduce((total, usage) => total + usage.seconds, 0) ?? 1;
   const usageColors = chroma
-    .scale([
-      theme.experimental.roles.success.fill.solid,
-      theme.experimental.roles.notice.fill.solid,
-    ])
+    .scale([theme.roles.success.fill.solid, theme.roles.notice.fill.solid])
     .mode("lch")
     .colors(validUsage?.length ?? 0);
   // The API returns a row for each app, even if the user didn't use it.
