+++ site/src/pages/UsersPage/UsersFilter.tsx
@@ -13,7 +13,7 @@ import {
   type UseFilterMenuOptions,
   useFilterMenu,
 } from "components/Filter/menu";
-import type { ThemeRole } from "theme/experimental";
+import type { ThemeRole } from "theme/roles";
 import { docs } from "utils/docs";
 
 const userFilterQuery = {
@@ -129,7 +129,7 @@ const StatusIndicator: FC<StatusIndicatorProps> = ({ option }) => {
         height: 8,
         width: 8,
         borderRadius: 4,
-        backgroundColor: theme.experimental.roles[option.color].fill.solid,
+        backgroundColor: theme.roles[option.color].fill.solid,
       }}
     />
   );
