+++ site/src/pages/WorkspacesPage/BatchUpdateConfirmation.tsx
@@ -477,7 +477,7 @@ const styles = {
   newVersion: (theme) => ({
     fontSize: 13,
     fontWeight: 500,
-    color: theme.experimental.roles.active.fill.solid,
+    color: theme.roles.active.fill.solid,
   }),
 
   message: {
