+++ site/src/pages/WorkspacesPage/LastUsed.tsx
@@ -40,16 +40,16 @@ export const LastUsed: FC<LastUsedProps> = ({ lastUsedAt }) => {
   );
 
   if (t.isAfter(now.subtract(1, "hour"))) {
-    circle = <Circle color={theme.experimental.roles.success.fill.solid} />;
+    circle = <Circle color={theme.roles.success.fill.solid} />;
     // Since the agent reports on a 10m interval,
     // the last_used_at can be inaccurate when recent.
     message = "Now";
   } else if (t.isAfter(now.subtract(3, "day"))) {
     circle = <Circle color={theme.palette.text.secondary} />;
   } else if (t.isAfter(now.subtract(1, "month"))) {
-    circle = <Circle color={theme.experimental.roles.warning.fill.solid} />;
+    circle = <Circle color={theme.roles.warning.fill.solid} />;
   } else if (t.isAfter(now.subtract(100, "year"))) {
-    circle = <Circle color={theme.experimental.roles.error.fill.solid} />;
+    circle = <Circle color={theme.roles.error.fill.solid} />;
   } else {
     message = "Never";
   }
