+++ site/src/theme/dark/index.ts
@@ -1,11 +1,13 @@
-import experimental from "./experimental";
-import monaco from "./monaco";
 import muiTheme from "./mui";
 import { forDarkThemes } from "../externalImages";
+import experimental from "./experimental";
+import monaco from "./monaco";
+import roles from "./roles";
 
 export default {
   ...muiTheme,
+  externalImages: forDarkThemes,
   experimental,
   monaco,
-  externalImages: forDarkThemes,
+  roles,
 };
