+++ site/src/theme/light/index.ts
@@ -1,11 +1,13 @@
-import experimental from "./experimental";
-import monaco from "./monaco";
 import muiTheme from "./mui";
 import { forLightThemes } from "../externalImages";
+import experimental from "./experimental";
+import monaco from "./monaco";
+import roles from "./roles";
 
 export default {
   ...muiTheme,
+  externalImages: forLightThemes,
   experimental,
   monaco,
-  externalImages: forLightThemes,
+  roles,
 };
