+++ site/src/utils/latency.ts
@@ -5,12 +5,12 @@ export const getLatencyColor = (theme: Theme, latency?: number) => {
     return theme.palette.text.secondary;
   }
 
-  let color = theme.experimental.roles.success.fill.solid;
+  let color = theme.roles.success.fill.solid;
 
   if (latency >= 150 && latency < 300) {
-    color = theme.experimental.roles.warning.fill.solid;
+    color = theme.roles.warning.fill.solid;
   } else if (latency >= 300) {
-    color = theme.experimental.roles.error.fill.solid;
+    color = theme.roles.error.fill.solid;
   }
   return color;
 };
